requirements.txt
asyncio==3.4.3
Babel==2.7.0
certifi==2019.11.28
Click==7.0
Flask==1.0.2
Flask-Babel==0.12.2
Flask-Cors==3.0.8
Flask-MySQL==1.4.0
future==0.18.2
gunicorn==19.9.0
iso8601==0.1.12
itsdangerous==1.1.0
Jinja2==2.10
kafka-python==1.4.6
MarkupSafe==1.1.1
modbus-tk==1.1.0
mysql-connector-python==8.0.16
paho-mqtt==1.5.0
PyMySQL==0.9.3
pyserial==3.4
pytz==2019.2
PyYAML==5.3
serial==0.0.97
six==1.12.0
Werkzeug==0.15.1
xmltodict==0.12.0
